Format document

+++ contracts/solc_0.8/defiai/FakeToken.sol
@@ -3,9 +3,9 @@ pragma solidity 0.8.4;
 import "../common/BEP20.sol";
 
 contract FakeToken is BEP20 {
-  constructor(string memory name, string memory symbol) BEP20(name, symbol) {}
+    constructor(string memory name, string memory symbol) BEP20(name, symbol) {}
 
-  function devMint(uint amount) external {
+    function devMint(uint256 amount) external {
         _mint(msg.sender, amount);
-  }
+    }
 }
